Move new stuff to partial classes
